+++ judge-stack/src/main/java/com/dwarfeng/judge/stack/handler/AnalysisFileFileOperateHandler.java
@@ -42,9 +42,10 @@ public interface AnalysisFileFileOperateHandler extends Handler {
      * 如果上传的分析结果文件文件存在，则进行替换。
      *
      * @param info 上传信息。
+     * @return 上传结果。
      * @throws HandlerException 处理器异常。
      */
-    void uploadFile(AnalysisFileFileUploadInfo info) throws HandlerException;
+    AnalysisFileFileUploadResult uploadFile(AnalysisFileFileUploadInfo info) throws HandlerException;
 
     /**
      * 通过流的方式上传分析结果文件。
@@ -61,7 +62,8 @@ public interface AnalysisFileFileOperateHandler extends Handler {
      * 即其中的 {@link AnalysisFileFileStreamUploadInfo#getContent()}。
      *
      * @param info 上传信息。
+     * @return 上传结果。
      * @throws HandlerException 处理器异常。
      */
-    void uploadFileStream(AnalysisFileFileStreamUploadInfo info) throws HandlerException;
+    AnalysisFileFileUploadResult uploadFileStream(AnalysisFileFileStreamUploadInfo info) throws HandlerException;
 }